refactor: APIコードをモジュール分割
680行の単一ファイル（api/index.py）を責務ごとに4ファイルへ分割:
- models.py: GeneFeature, GeneStructure クラスとPydanticモデル
- drawer.py: 描画定数とSVG描画ロジック
- parser.py: GFFパーサー
- index.py: FastAPIアプリとエンドポイント
